xrtx: Don't hardcode UART major
Currently thedriver hardcodes the UART major as 234. However, this is
VERY BAD (TM) as 234 is reserved for dynamic allocation BY THE KERNEL.
Trying to use 234 in this way in the driver will cause conflicts with
other dynamically major-allocated drivers in the kernel. Fortunately,
simply leaving out the `.major/.minor` field will cause the kernel
to dynamically allocate a proper major/minor number for the UART device.

Fixes: https://github.com/xtrx-sdr/images/issues/102
Signed-off-by: Keno Fischer <keno@juliacomputing.com>